Move state chain code to Shared lib. Add seperate Error.rs to shared lib also
