+++ shared/src/lib.rs
@@ -13,10 +13,11 @@ extern crate serde;
 extern crate serde_json;
 
 pub mod util;
+pub mod error;
 pub mod structs;
 pub mod state_chain;
 
-type Result<T> = std::result::Result<T, util::SharedLibError>;
+type Result<T> = std::result::Result<T, error::SharedLibError>;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct Root {
